Ignore description for default SG
Description for the default security group is hard-coded in OpenStack Neutron and cannot be
specified on creation (and cannot be updated as non-default SGs).
This patch changes the builder to ignore the description for the default SG in application,
in order not to destroy and recreate the default SG when the hard-coded description does not
match the one given in YAML configuration.